ArticleLuceneQueryBuilder implemented
