+++ component/src/main/resources/META-INF/components.txt
@@ -11,4 +11,5 @@ com.celements.blog.observation.listener.ClearBlogCacheListener
 com.celements.blog.observation.listener.ArticleDatesListener
 com.celements.blog.article.ArticleEngineLucene
 com.celements.blog.article.ArticleEngineHQL
+com.celements.blog.article.ArticleLuceneQueryBuilder
 com.celements.blog.migrator.ArticleNullDatesMigrator
